# contracts/internal/internal-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SuperAgentSystem Internal API
  version: v1
  summary: Launch Spec v1 frozen internal contracts with Memory System V3 single-tenant extension
servers:
  - url: https://internal.super-agent.local
components:
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
    Scope:
      name: X-Scope
      in: header
      required: true
      schema:
        type: string
    TraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Standard error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error_code, message, trace_id, retryable, details]
      properties:
        error_code:
          type: string
        message:
          type: string
        trace_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
...
    CleanupRequest:
      type: object
      required: [task_request_id, task_plan_id, task_step_id, cleanup_status, journal_cursor, capsule_id, dependency_state]
      properties:
        task_request_id:
          type: string
          format: uuid
        task_plan_id:
          type: string
          format: uuid
        task_step_id:
          type: string
          format: uuid
        cleanup_status:
          type: string
        journal_cursor:
          type: integer
        capsule_id:
          type: string
          format: uuid
        dependency_state:
          type: string
          enum: [DOWN, HALF-OPEN, UP]
    CleanupResponse:
      type: object
      required: [cleanup_status, drift_detected, scope_frozen, reconciliation_required]
      properties:
        cleanup_status:
          type: string
        drift_detected:
          type: boolean
        dlq_item_id:
          type: string
          format: uuid
        scope_frozen:
          type: boolean
        reconciliation_required:
          type: boolean
...
paths:
...
  /internal/cleanup/dispatch:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CleanupRequest"
      responses:
        "200":
          description: Cleanup dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanupResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...